refactor and doc
