+++ burn-autodiff/src/checkpoint/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod base;
+pub(crate) mod builder;
 pub(crate) mod state;
 #[cfg(test)]
-mod tests_ops;
+mod tests;
