+++ burn-autodiff/src/checkpoint/tmp
@@ -1,11 +1,6 @@
 TODOs
 
-- rename unsure stuff and might_need
-- might_need
-  - put inside memory bound
-  - allow stateless
-- maybe have only one attribute checkpointactions (the struct) in the ops prep
-- refactor the build algorithm
+- PROBLEM: retro_forwards are not well drained
 - remove tests, rename tests_ops, clean them
 - clean left out comments and prints
